full of errors

+++ IRF_project/IRF_project/Entities/Customers.cs
@@ -13,7 +13,7 @@ namespace IRF_project.Entities
             public Gender Gender { get; set; }
             public String Email { get; set; }
             public int Weight { get; set; }
-            public int Height { get; set; }
+            public float Height { get; set; }
             public String Lead_Source { get; set; }
